Adjust JavaDoc, refactor commands to validators
